+++ electron/ipc/store.ipc.js
@@ -1,6 +1,7 @@
 const { ipcMain } = require('electron');
 const storeService = require('../services/store.service');
 const { logError } = require('../services/logger.service');
+const updaterService = require('../services/updater.service');
 
 module.exports = function () {
 
@@ -15,10 +16,15 @@ module.exports = function () {
 
     ipcMain.handle("save-settings", async (_e, settings) => {
         try {
-            return await storeService.setSettings(settings);
+            storeService.setSettings(settings);
         } catch (err) {
             logError('IPC save-settings failed', err);
             throw err;
+        }
+        try {
+            updaterService.applyFromStoredSettings();
+        } catch (err) {
+            logError('IPC save-settings: apply updater preferences failed', err);
         }
     });