+++ src/app/features/workspace/titlebar/settings/settings.component.spec.ts
@@ -40,7 +40,9 @@ describe('SettingsComponent', () => {
     headers: {},
     ssl: { certificates: [] },
     dns: {},
-    proxy: {}
+    proxy: {},
+    logging: {},
+    updates: { allowPrerelease: false, allowDowngrade: false, targetRelease: 'latest' },
   };
 
   beforeEach(async () => {
@@ -67,6 +69,7 @@ describe('SettingsComponent', () => {
       'checkForUpdates',
       'downloadUpdate',
       'installUpdate',
+      'listUpdaterReleases',
     ]) as jasmine.SpyObj<UpdateService> & { statusStream: unknown };
     (updateServiceSpy as { statusStream: unknown }).statusStream = of({
       state: 'idle',
@@ -74,6 +77,7 @@ describe('SettingsComponent', () => {
       supported: false,
       info: null,
     } as any);
+    updateServiceSpy.listUpdaterReleases.and.resolveTo([]);
 
     environmentsServiceSpy = jasmine.createSpyObj('EnvironmentsService', [
       'loadEnvironments',